+++ src/main/java/Command/MusicModule/SkipCommand.java
@@ -13,6 +13,7 @@ import Resource.Info;
 import Setting.Prefix;
 import java.awt.Color;
 import java.time.Instant;
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.MessageEmbed;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
@@ -21,9 +22,11 @@ import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
  * @author Alien Ideology <alien.ideology at alien.org>
  */
 public class SkipCommand implements Command {
-    public final static  String HELP = "This command is for skipping the current song.\n"
-                                     + "Command Usage: `"+ Prefix.getDefaultPrefix() +"skip`\n"
-                                     + "Parameter: `-h | null`";
+    public final static String HELP = "This command is for skipping the current song.\n"
+                                    + "Command Usage: `"+ Prefix.getDefaultPrefix() +"skip`\n"
+                                    + "Parameter: `-h | -f | null`\n"
+                                    + "null: Vote to skipp current song.\n"
+                                    + "-f: Force skipp current song. Requires Administrator or Manage Channel Permission.";
     
     @Override
     public boolean called(String[] args, MessageReceivedEvent e) {
@@ -64,7 +67,19 @@ public class SkipCommand implements Command {
         
         else if("-f".equals(args[0]))
         {
-            Music.skip(e, 0, true);
+            if(e.getMember().isOwner() || 
+                    e.getMember().hasPermission(Permission.ADMINISTRATOR) || 
+                    e.getMember().hasPermission(Permission.MANAGE_CHANNEL) || 
+                    Info.D_ID.equals(e.getAuthor().getId()))
+            {
+                Music.skip(e, 0, true);
+                e.getChannel().sendMessage(Emoji.success + " Track skipped.").queue();
+            }
+            else
+            {
+                e.getChannel().sendMessage(Emoji.error + " Only server owner and members with \n"
+                        + "`Administrator` or `Manage Channel` permission can force skip a song.").queue();
+            }
         }
         
         else if(args[0].length() == 1 && Character.isDigit(args[0].charAt(0)))