+++ src/main/java/Resource/Info.java
@@ -39,5 +39,7 @@ public class Info {
     //Link
     public static String L_MUSIC_HUB = "https://discord.gg/UMCqtZN";
     public static final String LYRICSURL = "https://genius.com/";
-
+    
+    //Bot Developer ID
+    public static String D_ID = "248214880379863041";
 }